+++ src/UI/Data/Service/ServiceStateHelper.ts
@@ -1,20 +1,20 @@
-import { RemoteData, ResourceModel, StateHelper } from "@/Core";
+import { Query, RemoteData, StateHelper } from "@/Core";
 import { Store, useStoreState } from "@/UI/Store";
 import { isEqual } from "lodash";
 
-type Data = RemoteData.Type<string, ResourceModel[]>;
+type Data = RemoteData.Type<Query.Error<"Service">, Query.Data<"Service">>;
 
-export class ResourcesStateHelper
-  implements StateHelper<string, ResourceModel[]> {
+export class ServiceStateHelper
+  implements StateHelper<Query.Error<"Service">, Query.Data<"Service">> {
   constructor(private readonly store: Store) {}
 
   set(id: string, value: Data): void {
-    this.store.dispatch.resources.setData({ id, value });
+    this.store.dispatch.services2.setData({ id, value });
   }
 
   getHooked(id: string): Data {
     return useStoreState((state) => {
-      return this.enforce(state.resources.byId[id]);
+      return this.enforce(state.services2.byId[id]);
     }, isEqual);
   }
 
@@ -24,6 +24,6 @@ export class ResourcesStateHelper
   }
 
   getOnce(id: string): Data {
-    return this.enforce(this.store.getState().resources.byId[id]);
+    return this.enforce(this.store.getState().services2.byId[id]);
   }
 }