+++ src/app/routes.tsx
@@ -8,8 +8,8 @@ import {
   useLastLocation,
 } from "react-router-last-location";
 import { ServiceCatalog } from "./ServiceCatalog/ServiceCatalog";
-import { ServiceInventory } from "./ServiceInventory/ServiceInventory";
 import { CreateInstancePage } from "@/UI/ServiceInstanceForm";
+import { ServiceInventoryWithProvider } from "./ServiceInventory/ServiceInventory";
 
 let routeFocusTimer: number;
 
@@ -45,7 +45,7 @@ const routes: IAppRouteGroup[] = [
         title: "Service Catalog",
       },
       {
-        component: ServiceInventory,
+        component: ServiceInventoryWithProvider,
         exact: true,
         hideOnSideBar: true,
         icon: null,